% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIT_DMN_adj_p.R
\name{AIT_DMN_adj_p}
\alias{AIT_DMN_adj_p}
\title{An adaptive independence test for Dirichlet-multinomial modeled microbiome data after covariates adjustment}
\usage{
AIT_DMN_adj_p(x, y, z, lambda, permutation = 1000, slice = FALSE)
}
\arguments{
\item{x}{a vector of bacterial counts follows dirichlet-multinomial distribution.}

\item{y}{a continuous or categorical outcome vector. We use MLE to estimate coefficent, thus our method prefers
a pre-sliced (categorical) outcome so that we have enough samples to estimate parameters.}

\item{z}{the covariates.}

\item{lambda}{the panalty for the number of slices.}

\item{permutation}{permutation time for the test (default 1000). It is recommended to permutate
100 times to reduce the amount of calculation while \code{y} is a continuous variable.}

\item{slice}{a bool variable to determine whether return slicing scheme (default FALSE).}
}
\value{
results of the test: \code{statistics} and \code{p_value} (if \code{slice=TRUE}, also return \code{slicing}).
}
\description{
An adaptive independence test based on the Dirichlet-multinomial distribution. It adjusts covariate,
then tests the independence between a vector of bacterial counts and a continuous variable or a
categorical variable with many levels.
}
\examples{

##################################### a simulated example ###################################
rm(list=ls())
set.seed(2019)
p <- 6  # number of taxa
d <- 1  # number of covariates
samples <- 100
seq.depth=500
mean.proportion=rnorm(p,1,1)
groups.samples<-c(10,20,30,40)
gamma=t(matrix(sort(rep(0:3,p)),p,4))
coefficent=matrix(rnorm(p),1,p)
covariates=matrix(rnorm(samples),samples,1)
x.dc=sim_x_c(seq.depth,mean.proportion,gamma,groups.samples,coefficent,covariates)
y=sort(rep(1:10,10))/2
y.dc=y
# lambda_da=lambda_selection("DMN",x.dc,y.dc,permutation=100,covariates=covariates) #2.75
lambda_da=2.75
AIT_DMN_adj_p(x.dc,y.dc,covariates,lambda_da,permutation=100,slice=TRUE)

}
\seealso{
\code{\link{sim_x_c}}
}
