% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIT_MN_p.R
\name{AIT_MN_p}
\alias{AIT_MN_p}
\title{An adaptive independence test for multinomial distributed microbiome data}
\usage{
AIT_MN_p(x, y, lambda, permutation = 1000, slice = FALSE)
}
\arguments{
\item{x}{a vector of bacterial counts follows multinomial distribution.}

\item{y}{a continuous or categorical outcome vector.}

\item{lambda}{the penalty for the number of slices.}

\item{permutation}{permutation time for the test (default 1000). It is recommended to permutate
100 times to reduce the amount of calculation while \code{y} is a continuous variable.}

\item{slice}{a bool variable to determine whether return slicing scheme (default FALSE).}
}
\value{
results of the test: \code{statistics} and \code{p_value} (if \code{slice=TRUE}, also return \code{slicing}).
}
\description{
An adaptive independence test based on the multinomial distribution. It tests the independence
between a vector of bacterial counts and a continuous variable or a categorical variable with
many levels.
}
\examples{
set.seed(2019)
data(x.m)
y=sort(rep(1:10,10))/2
y.m=y
theta=theta_calculate(x.m)
lambda_m=lambda_selection("MN",x.m,y.m,1000)
AIT_MN_p(x.m,y.m,lambda_m,1000,slice = TRUE)
lambda_d=lambda_selection("DMN",x.m,y.m,1000)
AIT_DMN_p(x.m,y.m,lambda_d,1000,slice = TRUE)

y.m=sort(runif(100, min=0.01, max=1))
lambda_m=lambda_selection("MN",x.m,y.m,100)
AIT_MN_p(x.m,y.m,lambda_m,1000,slice = TRUE)
}
