% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar.R
\name{Bar_server}
\alias{Bar_server}
\title{Bar_server Server Module}
\usage{
Bar_server(input, output, session)
}
\arguments{
\item{input}{Standard shiny server arguments}

\item{output}{Standard shiny server arguments}

\item{session}{Standard shiny server arguments}
}
\description{
Server logic for merging files Bar_server
}
