% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{boxplot_server}
\alias{boxplot_server}
\title{boxplot_server Module}
\usage{
boxplot_server(input, output, session)
}
\arguments{
\item{input}{Standard shiny server arguments}

\item{output}{Standard shiny server arguments}

\item{session}{Standard shiny server arguments}
}
\description{
Server logic for boxplot_server
}
