% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{boxplot_ui}
\alias{boxplot_ui}
\title{boxplot_ui}
\usage{
boxplot_ui(id)
}
\arguments{
\item{id}{A unique identifier for the Shiny namespace.}
}
\description{
boxplot UI Module
}
\details{
boxplot UI Module
}
