% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCA.R
\name{CCA_server}
\alias{CCA_server}
\title{CCA_server Server Module}
\usage{
CCA_server(input, output, session)
}
\arguments{
\item{input}{Standard shiny server arguments}

\item{output}{Standard shiny server arguments}

\item{session}{Standard shiny server arguments}
}
\description{
Server logic for CCA_server
}
