% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cor.R
\name{cor_server}
\alias{cor_server}
\title{Server for Correlation Analysis}
\usage{
cor_server(input, output, session)
}
\arguments{
\item{input}{Input object}

\item{output}{Output object}

\item{session}{Shiny session object.}
}
\description{
Handles the server-side logic for correlation analysis
}
