% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corxy.R
\name{corxy_server}
\alias{corxy_server}
\title{Calculate the correlation between two sets of data server}
\usage{
corxy_server(input, output, session)
}
\arguments{
\item{input}{A unique identifier for the Shiny namespace.}

\item{output}{A unique identifier for the Shiny namespace.}

\item{session}{A unique identifier for the Shiny namespace.}
}
\description{
corxy UI Module
}
