% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corxy.R
\name{corxy_ui}
\alias{corxy_ui}
\title{Calculate the correlation between two sets of data}
\usage{
corxy_ui(id)
}
\arguments{
\item{id}{A unique identifier for the Shiny namespace.}
}
\description{
corxy UI Module
}
