% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\docType{data}
\name{custom_css}
\alias{custom_css}
\title{Custom CSS}
\format{
An object of class \code{character} of length 1.
}
\usage{
custom_css
}
\description{
Add custom CSS to adjust the dropdown menu and toggle styles.
}
\keyword{datasets}
