% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\docType{data}
\name{customLogo}
\alias{customLogo}
\title{Custom Logo}
\format{
An object of class \code{html} (inherits from \code{character}) of length 1.
}
\usage{
customLogo
}
\description{
Create a custom logo for the dashboard.
}
\keyword{datasets}
