% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transform_module.R
\name{dataTransform_server}
\alias{dataTransform_server}
\title{Data Transformation and Normalization Server Module}
\usage{
dataTransform_server(input, output, session)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}
}
\value{
None
}
\description{
Data Transformation and Normalization Server Module
}
