% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataTransform.R
\name{dataTransform_ui}
\alias{dataTransform_ui}
\title{dataTransform UI Module}
\usage{
dataTransform_ui(id)
}
\arguments{
\item{id}{A unique identifier for the Shiny namespace.}
}
\description{
Merge File UI Module
}
