% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Describe.R
\name{describe_server}
\alias{describe_server}
\title{Server for Descriptive Statistics}
\usage{
describe_server(input, output, session)
}
\arguments{
\item{input}{Input object}

\item{output}{Output object}

\item{session}{Shiny session object.}
}
\description{
Handles the server-side logic for descriptive statistics analysis using psych::describe
}
