% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Describe.R
\name{describe_ui}
\alias{describe_ui}
\title{UI for Descriptive Statistics}
\usage{
describe_ui(id)
}
\arguments{
\item{id}{Namespace ID}
}
\description{
Creates the UI for descriptive statistics analysis using psych::describe
}
