% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq2.R
\name{deseq2_server}
\alias{deseq2_server}
\title{DESeq2 Server Module}
\arguments{
\item{input}{Standard shiny server arguments}

\item{output}{Standard shiny server arguments}

\item{session}{Standard shiny server arguments}
}
\description{
Server logic for DESeq2 analysis
}
