% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_bubble.R
\name{enrichment_bubble_server}
\alias{enrichment_bubble_server}
\title{title enrichment_bubble_server}
\arguments{
\item{id}{enrichment_bubble_server id}

\item{input}{description}

\item{output}{description}

\item{session}{description}
}
\description{
enrichment_bubble_server
}
