% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_bubble.R
\name{enrichment_bubble_ui}
\alias{enrichment_bubble_ui}
\title{enrichment_bubble}
\usage{
enrichment_bubble_ui(id)
}
\arguments{
\item{id}{A unique identifier for the Shiny namespace.}
}
\description{
enrichment_bubble UI Module
}
\details{
enrichment_bubble UI Module
}
