% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexible_tools.R
\name{flexible_tools_ui}
\alias{flexible_tools_ui}
\title{flexible_tools_ui}
\usage{
flexible_tools_ui(id)
}
\arguments{
\item{id}{A unique identifier for the Shiny namespace.}
}
\description{
Flexible Tools User Interface
}
\details{
Creates a user interface for flexible data export tools in a Shiny application.
This panel allows users to select data export formats and specify output filenames.
}
