% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO_bar_class.R
\name{GO_bar_class_server}
\alias{GO_bar_class_server}
\title{GO Bar Plot Server}
\usage{
GO_bar_class_server(input, output, session)
}
\description{
Handles the server-side logic for the GO Bar Plot application, including reading uploaded data,
generating the plot based on user inputs, and enabling PDF downloads with specified dimensions.
}
