% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO_bar_class.R
\name{GO_bar_class_ui}
\alias{GO_bar_class_ui}
\title{GO_bar_class_ui}
\usage{
GO_bar_class_ui(id)
}
\arguments{
\item{id}{A unique identifier for the Shiny namespace.}
}
\description{
GO_bar_class UI Module
}
\details{
GO_bar_class UI Module
}
