% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grouping_sorting.R
\name{Grouping_sorting_ui}
\alias{Grouping_sorting_ui}
\title{Grouping_sorting UI Module}
\usage{
Grouping_sorting_ui(id)
}
\arguments{
\item{id}{A unique identifier for the Shiny namespace.}
}
\description{
Merge File UI Module
}
