% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grouping_statistics.R
\name{Grouping_statistics_server}
\alias{Grouping_statistics_server}
\title{Grouping Statistics Server Function}
\usage{
Grouping_statistics_server(input, output, session)
}
\arguments{
\item{input}{Shiny input object, used to access input values from the UI.}

\item{output}{Shiny output object, used to send output values to the UI.}

\item{session}{Shiny session object, used to manage session-level information.}
}
\description{
This function defines the server logic for grouping statistics in a Shiny application.
}
