% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grouping_statistics.R
\name{Grouping_statistics_ui}
\alias{Grouping_statistics_ui}
\title{Grouping_statistics UI Module}
\usage{
Grouping_statistics_ui(id)
}
\arguments{
\item{id}{A unique identifier for the Shiny namespace.}
}
\description{
Grouping_statistics UI Module
}
