% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{histogram_server}
\alias{histogram_server}
\title{histogram_server}
\arguments{
\item{input}{description}

\item{output}{description}

\item{session}{description}
}
\description{
histogram_server Module
}
\details{
histogram UI Module
}
