% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{histogram_ui}
\alias{histogram_ui}
\title{histogram_ui}
\usage{
histogram_ui(id)
}
\arguments{
\item{id}{A unique identifier for the Shiny namespace.}
}
\description{
histogram UI Module
}
\details{
histogram UI Module
}
