% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans.R
\name{KmeansR_Ait}
\alias{KmeansR_Ait}
\title{title}
\arguments{
\item{data}{A data frame or matrix containing the data to be clustered. Rows represent observations,
and columns represent features.}

\item{centers}{An integer specifying the number of clusters to generate in the K-means analysis.}
}
\description{
title
}
