% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_selection.R
\name{lambda_selection}
\alias{lambda_selection}
\title{Select a proper lambda for adaptive independence test.}
\usage{
lambda_selection(method = "DMN", x, y, permutation = 100, sig = 0.05,
  covariates = NULL)
}
\arguments{
\item{method}{two option: 'MN' or 'DMN' for multinomial and Dirichlet-multinomial model seperately (default 'DMN').}

\item{x}{a matrix represents microbiomal composition.}

\item{y}{a continuous or categorical covariate vector.}

\item{permutation}{permutation time for the test (default 1000). It is recommended to permutate 100 times to reduce
the amount of calculation when \code{y} is continuous.}

\item{sig}{significance level, that's type I error.}

\item{covariates}{covariates for microbiome composition.}
}
\value{
a proper penalty lambda for the the adaptive independence test.
}
\description{
Select a proper lambda to control type I error for the independence test.
}
\examples{

data(x.m)
data(x.d)
y=sort(rep(1:10,10))/2
y.d=y.m=y
lambda_m=lambda_selection("MN",x.m,y.m,100)
lambda_d=lambda_selection("DMN",x.d,y.d,100)
}
\seealso{
\code{\link{x.m}},\code{\link{x.d}}
}
