% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Long_Wide_Data_T.R
\name{Long_Wide_Data_T_ui}
\alias{Long_Wide_Data_T_ui}
\title{dataTransform UI Module}
\usage{
Long_Wide_Data_T_ui(id)
}
\arguments{
\item{id}{A unique identifier for the Shiny namespace.}
}
\description{
Merge File UI Module
}
