% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_file.R
\name{merge_file_server}
\alias{merge_file_server}
\title{Merge File Server Module}
\usage{
merge_file_server(input, output, session)
}
\arguments{
\item{input}{Standard shiny server arguments}

\item{output}{Standard shiny server arguments}

\item{session}{Standard shiny server arguments}
}
\description{
Server logic for merging files
}
