% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_file.R
\name{merge_file_ui}
\alias{merge_file_ui}
\title{merge_file}
\usage{
merge_file_ui(id)
}
\arguments{
\item{id}{A unique identifier for the Shiny namespace.}
}
\description{
Merge File UI Module
}
\details{
Merge File UI Module
}
