% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mfuzz.R
\name{mfuzz_server}
\alias{mfuzz_server}
\title{Server logic for Mfuzz Clustering}
\usage{
mfuzz_server(input, output, session)
}
\arguments{
\item{id}{A unique identifier for the module.}
}
\value{
A Shiny server function that handles clustering and data processing.
}
\description{
Defines the server logic for the Mfuzz clustering module in a Shiny application.
This includes reading the input file, preprocessing the data, running the Mfuzz clustering algorithm,
and providing the results for download and visualization.
}
