% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mfuzz.R
\name{mfuzz_ui}
\alias{mfuzz_ui}
\title{Application User Interface for Mfuzz Clustering}
\arguments{
\item{id}{A unique identifier for the module.}
}
\value{
A Shiny UI object for the Mfuzz clustering module.
}
\description{
Defines the user interface for the Mfuzz clustering module in a Shiny application.
}
