% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPLS_DA.R
\name{OPLS_DA_server}
\alias{OPLS_DA_server}
\title{OPLS_DA_server Server Module}
\arguments{
\item{input}{Standard shiny server arguments}

\item{output}{Standard shiny server arguments}

\item{session}{Standard shiny server arguments}
}
\description{
Server logic for OPLS_DA_server
}
