% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{pca_server}
\alias{pca_server}
\title{PCA Server Module}
\usage{
pca_server(input, output, session)
}
\arguments{
\item{input, }{output, session Standard Shiny server arguments.}
}
\value{
None. This function is used for its side effects.
}
\description{
This module handles the server-side logic for the Principal Component Analysis (PCA) tab.
It includes reactive expressions for handling file inputs, generating the PCA plot, and 
allowing users to download the resulting plot and data.
}
