% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{pca_ui}
\alias{pca_ui}
\title{PCA UI Module}
\usage{
pca_ui(id)
}
\arguments{
\item{id}{The module's unique identifier.}
}
\value{
A Shiny UI layout for the PCA module.
}
\description{
This module creates the UI for the Principal Component Analysis (PCA) tab.
It includes file inputs for uploading the matrix and group files, color selector inputs, 
and controls for downloading the plot and data.
}
