% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pie.R
\name{pie_server}
\alias{pie_server}
\title{pie_server}
\arguments{
\item{input}{Standard shiny server arguments}

\item{output}{Standard shiny server arguments}

\item{session}{Standard shiny server arguments}
}
\description{
Server logic for pie
}
\details{
pie_server Module
}
