% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomNum.R
\name{randomNum_server}
\alias{randomNum_server}
\title{randomNum_server}
\usage{
randomNum_server(input, output, session)
}
\arguments{
\item{input}{description}

\item{output}{description}

\item{session}{description}
}
\description{
A short description...
}
