% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relabel.R
\name{relabel}
\alias{relabel}
\title{Reassigning values of a vector}
\usage{
relabel(y)
}
\arguments{
\item{y}{a continuous or categorical vector.}
}
\value{
an integer vector with values range from 1 to k (k > 1).
}
\description{
Reassigning values of a continuous or categorical vector. It is used to generate legal
value before applying independence test.
}
\examples{
n <- 10
y <- c(rep("G3", n),rep("G2", n), rep("G1", n/2))
y.r <- relabel(y)
}
