% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse_complement.R
\name{reverse_complement_ui}
\alias{reverse_complement_ui}
\title{reverse_complement_ui}
\usage{
reverse_complement_ui(id)
}
\arguments{
\item{id}{A unique identifier for the Shiny namespace.}
}
\description{
reverse_complement UI Module
}
\details{
reverse_complement UI Module
}
