% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgePlot.R
\name{ridgePlot_server}
\alias{ridgePlot_server}
\title{ridgePlot_server}
\usage{
ridgePlot_server(input, output, session)
}
\arguments{
\item{input}{description}

\item{output}{description}

\item{session}{description}
}
\description{
ridgePlot_server
}
