% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgePlot.R
\name{ridgePlot_ui}
\alias{ridgePlot_ui}
\title{ridgePlot_ui}
\usage{
ridgePlot_ui(id)
}
\arguments{
\item{id}{A unique identifier for the Shiny namespace.}
}
\description{
ridgePlot UI Module
}
\details{
ridgePlot UI Module
}
