% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\docType{data}
\name{run_Ait}
\alias{run_Ait}
\alias{folder_path}
\title{Run Ait Application}
\format{
An object of class \code{character} of length 1.
}
\usage{
folder_path

run_Ait(
  onStart = NULL,
  options = list(),
  enableBookmarking = NULL,
  uiPattern = "/",
  maxRequestSize = 100,
  ...
)
}
\arguments{
\item{onStart}{A function that will be run before the app is started.}

\item{options}{A list of options to be passed to shiny::shinyApp.}

\item{enableBookmarking}{Enable bookmarking for the app.}

\item{uiPattern}{A URL pattern to match Shiny requests.}

\item{maxRequestSize}{Maximum request size (in MB) for file uploads.}

\item{...}{Additional arguments to be passed to golem options.}
}
\description{
Run the Ait Shiny application.
}
\keyword{datasets}
