% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqLogo.R
\name{seqLogo_server}
\alias{seqLogo_server}
\title{seqLogo_server}
\usage{
seqLogo_server(input, output, session)
}
\arguments{
\item{input}{Shiny server input}

\item{output}{Shiny server output}

\item{session}{Shiny session object}
}
\description{
Server logic for DNA Sequence Logo Module
}
\details{
Handles file input, rendering of the sequence logo plot, and downloading of the plot.
}
