% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqLogo.R
\name{seqLogo_ui}
\alias{seqLogo_ui}
\title{seqLogo ui Module}
\usage{
seqLogo_ui(id, title = "Sequence Logo")
}
\arguments{
\item{id}{A unique identifier for the Shiny namespace.}

\item{title}{The title of the sequence logo UI, default is "Sequence Logo".}
}
\description{
seqLogo UI Module
}
\details{
seqLogo UI Module
}
