% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_extract.R
\name{sequence_extract_ui}
\alias{sequence_extract_ui}
\title{sequence_extract_ui}
\usage{
sequence_extract_ui(id)
}
\arguments{
\item{id}{A unique identifier for the Shiny namespace.}
}
\description{
sequence_extract UI Module
}
\details{
sequence_extract UI Module
}
