% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_x_c.R
\name{sim_x_c}
\alias{sim_x_c}
\title{Simulate microbiome composition data with covariates}
\usage{
sim_x_c(seq_depth, mean_proportion, gamma, groups_samples, coefficent,
  covariates)
}
\arguments{
\item{seq_depth}{sequencing depth for all samples. It is consistent for all samples.}

\item{mean_proportion}{a vector consisted with taxa propotion parameter \eqn{\mu} for each taxa.}

\item{gamma}{coefficient matrix \eqn{\gamma}, each row corresponds to each group, and each column is for taxa.}

\item{groups_samples}{a vector with number of samples in each group.}

\item{coefficent}{coefficient \eqn{c} of \code{covariates} for each taxon .}

\item{covariates}{covariates matrix \eqn{z} for microbiome composition.}
}
\value{
a microbiome composition matrix.
}
\description{
Simulate Dirichlet-multinomial distributed microbiome data subjected to covariates.
}
\examples{

rm(list=ls())
set.seed(2019)
p <- 6  # number of taxa
d <- 1  # number of covariates
samples <- 100
seq.depth=500
mean.proportion=rnorm(p,1,1)
groups.samples<-c(10,20,30,40)
gamma=t(matrix(sort(rep(0:3,p)),p,4))
#gamma=t(matrix(0,p,4))
coefficent=matrix(rnorm(p),1,p)
covariates=matrix(rnorm(samples),samples,1)
y=sort(rep(1:10,10))
x.dc=sim_x_c(seq.depth,mean.proportion,gamma,groups.samples,coefficent,covariates)


}
