% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svg_ui.R
\name{svg_server}
\alias{svg_server}
\title{Server for SVG conversion}
\usage{
svg_server(input, output, session)
}
\arguments{
\item{input}{Input object}

\item{output}{Output object}

\item{session}{Shiny session object}
}
\description{
Handles the server-side logic for SVG conversion
}
