\name{taxa_family}
\alias{taxa_family}
\docType{data}
\title{
Taxa counts at family level for samples.
}
\description{

Yatsunenko et al. (2012) conducted a study to investigate whether there exist taxa abundance differences across age and geography, where fecal specimens were collected from 528 healthy infants, children, and adults living in Amazonas (Venezuela), rural Malawi, and USA metropolitan areas. We combined the OTUs at the family level (242 families) and set it as dataset \code{taxa_family}.

}
\usage{data(taxa_family)}
\format{
The data \code{taxa_family} has 528 rows represent individuals, and 242 columns show taxa at family level.
}

\examples{
data(taxa_family)
}
\keyword{datasets}
\source{
Yatsunenko, T., Rey, F. E., Manary, M. J., Trehan, I., Dominguez-Bello, M. G., Contreras, M., Magris, M., Hidalgo, G., Baldassano, R. N., Anokhin, A. P., et al. (2012). Human gut microbiome viewed across age and geography. Nature, 486(7402), 222–227.
}
