% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theta_calculate.R
\name{theta_calculate}
\alias{theta_calculate}
\title{Calculate overdispersion for microbiomal data}
\usage{
theta_calculate(x)
}
\arguments{
\item{x}{a matrix follows multinomial or Dirichlet_multinomial distribution.}
}
\value{
overall moment estimator for overdispersion \eqn{\theta}.
}
\description{
Calculate overdispersion \eqn{\theta} for microbiomal data. If \eqn{\theta>0}, we can assume microbiome follows
Dirichlet-multinomial distribution; otherwise, it follows multinomial distribution.
}
\examples{
data(x.d)
data(x.m)
theta_calculate(x.d)
theta_calculate(x.m)
}
\seealso{
\code{\link{x.d}}, \code{\link{x.m}}
}
