% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upset.R
\name{upset_server}
\alias{upset_server}
\title{upset_server}
\arguments{
\item{input}{description}

\item{output}{description}

\item{session}{description}
}
\description{
upset_server Module
}
\details{
upset_server Module
}
