% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upset.R
\name{upset_ui}
\alias{upset_ui}
\title{User Interface for UpSetR Visualization}
\usage{
upset_ui(id)
}
\arguments{
\item{id}{A unique identifier for this UI component.}
}
\value{
A `shiny::tabPanel` object containing the UI elements for UpSetR visualization.
}
\description{
This function defines the UI components for the UpSetR visualization tab,
including file input, numeric inputs for plot dimensions, and download buttons.
}
