% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Venn.R
\name{Venn_server}
\alias{Venn_server}
\title{Venn_server}
\usage{
Venn_server(input, output, session)
}
\arguments{
\item{input}{Standard shiny server arguments}

\item{output}{Standard shiny server arguments}

\item{session}{Standard shiny server arguments}
}
\description{
Server logic for Venn_server
}
\details{
Venn_server Server Module
}
