% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VolcanoPlot.R
\name{VolcanoPlot_server}
\alias{VolcanoPlot_server}
\title{VolcanoPlot_server Module}
\arguments{
\item{input}{Standard shiny server arguments}

\item{output}{Standard shiny server arguments}

\item{session}{Standard shiny server arguments}
}
\description{
Server logic for VolcanoPlot_server
}
