% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VolcanoPlot.R
\name{VolcanoPlot_ui}
\alias{VolcanoPlot_ui}
\title{VolcanoPlot_ui}
\usage{
VolcanoPlot_ui(id)
}
\arguments{
\item{id}{A unique identifier for the Shiny namespace.}
}
\description{
VolcanoPlot UI Module
}
\details{
VolcanoPlot UI Module
}
