% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordcloud2.R
\name{wordcloud2_server}
\alias{wordcloud2_server}
\title{Server Logic for Word Cloud Module}
\usage{
wordcloud2_server(input, output, session)
}
\arguments{
\item{input}{Shiny server input.}

\item{output}{Shiny server output.}

\item{session}{Shiny server session.}
}
\value{
Does not return anything; it registers reactive values and outputs for a Shiny module.
}
\description{
This function contains the server-side logic for the Word Cloud module. It processes the uploaded
CSV file, generates a word cloud, and handles the download of the word cloud in PNG or HTML format.
}
