% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordcloud2.R
\name{wordcloud2_ui}
\alias{wordcloud2_ui}
\title{User Interface for Word Cloud Module}
\usage{
wordcloud2_ui(id)
}
\arguments{
\item{id}{A unique identifier for the UI elements to handle namespaces in Shiny modules.}
}
\value{
Returns a tab panel that includes a sidebar for inputs and a main panel for word cloud output.
}
\description{
This function creates a user interface for the Word Cloud module in a Shiny application.
It allows users to upload a CSV file and set parameters for generating a word cloud.
}
