% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCA.R
\name{CCA_server}
\alias{CCA_server}
\title{Server Function for CCA Module}
\usage{
CCA_server(input, output, session)
}
\description{
Handles server-side logic for the CCA (Canonical Correlation Analysis) module.
}
