% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transform_module.R
\name{dataTransform_ui}
\alias{dataTransform_ui}
\title{Data Transformation and Normalization UI Module}
\usage{
dataTransform_ui(id)
}
\arguments{
\item{id}{Module ID}
}
\value{
A fluidPage containing the UI elements for data transformation and normalization
}
\description{
Data Transformation and Normalization UI Module
}
