% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO_bar_class.R
\name{GO_bar_class_server}
\alias{GO_bar_class_server}
\title{GO_bar_class_server}
\arguments{
\item{id}{A unique identifier for the Shiny namespace.}

\item{input}{description}

\item{output}{description}

\item{session}{description}
}
\description{
GO_bar_class_server Module
}
\details{
GO_bar_class_server Module
}
