% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO_bar_class.R
\name{GO_bar_class_ui}
\alias{GO_bar_class_ui}
\title{GO Bar Plot UI}
\usage{
GO_bar_class_ui(id)
}
\description{
Creates the user interface for the GO Bar Plot application, allowing users to upload a CSV file,
select colors for different categories, and customize plot labels and dimensions.
}
