% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfuzz.R
\name{mfuzz_server}
\alias{mfuzz_server}
\title{Server logic for Mfuzz Clustering}
\arguments{
\item{input}{A unique identifier for the module.}

\item{output}{description}

\item{session}{description}
}
\value{
A Shiny server function that handles clustering and data processing.
}
\description{
Defines the server logic for the Mfuzz clustering module in a Shiny application.
This includes reading the input file, preprocessing the data, running the Mfuzz clustering algorithm,
and providing the results for download and visualization.
}
