% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPLS_DA.R
\name{OPLS_DA_ui}
\alias{OPLS_DA_ui}
\title{OPLS_DA UI Module}
\usage{
OPLS_DA_ui(id)
}
\arguments{
\item{id}{A unique identifier for the module.}
}
\description{
UI for displaying OPLS_DA (Orthogonal Projections to Latent Structures Discriminant Analysis)
}
