% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca_server}
\alias{pca_server}
\title{pca_server}
\arguments{
\item{input}{A unique identifier for the Shiny namespace.}

\item{output}{A unique identifier for the Shiny namespace.}

\item{session}{A unique identifier for the Shiny namespace.}
}
\description{
pca pca_server
}
\details{
pca UI pca_server
}
