% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_logo.R
\name{seqLogo_ui}
\alias{seqLogo_ui}
\title{UI for DNA Sequence Logo Module}
\usage{
seqLogo_ui(id, title = "Sequence Logo")
}
\arguments{
\item{id}{Unique identifier for the Shiny module}

\item{title}{Title of the tab panel}
}
\value{
Returns a tab panel with user interface elements for DNA sequence logo generation.
}
\description{
UI for DNA Sequence Logo Module
}
