% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upset.R
\name{upset_server}
\alias{upset_server}
\title{Server logic for UpSetR Visualization}
\usage{
upset_server(input, output, session)
}
\arguments{
\item{input}{A list of input values from the UI.}

\item{output}{A list of output values to be rendered in the UI.}

\item{session}{The Shiny session object.}
}
\description{
This function contains the server logic for rendering the UpSetR plot and 
handling file uploads and downloads. It processes the uploaded CSV data, 
generates the UpSetR plot, and provides options to download both the plot 
and the processed data.
}
